refactor(chat): drop bridge-router architecture, ship Telegram long-poll only
The standard ``python-telegram-bot`` long-poll driver is now the only
Telegram driver: bot API token from ``@BotFather``, chat id, ``pip
install bernstein[telegram]``, no external services. The earlier
optional bridge-router-service architecture pointed at a deployment
public users could not stand up themselves, so it has been removed.

Files removed:
- ``drivers/_legacy_telegram.py`` (content promoted to ``drivers/telegram.py``)
- ``tests/unit/test_chat_telegram_bridge.py`` (no more bridge to test)

Files simplified:
- ``drivers/telegram.py`` -- canonical long-poll driver, docstring rewritten
- ``core/chat/__init__.py`` -- ``load_driver("telegram")`` returns the
  long-poll driver directly; no env-flag selection
- ``notifications/sinks/telegram.py`` -- single ``deliver`` path through
  the long-poll bridge; ``bridge=`` (live) or ``token=`` config keys
- ``pyproject.toml`` -- ``[telegram]`` extra comment trimmed
- ``CHANGELOG.md`` -- Unreleased entry rewritten
- ``tests/unit/test_chat_drivers.py`` -- imports updated

Test plan: 90 chat + notifications unit tests pass; ruff check +
ruff format clean.
