+++ src/bernstein/core/chat/drivers/telegram.py
@@ -1,39 +1,40 @@
-"""Telegram driver -- thin shim over the ``thisnotabot-bridge`` SDK.
+"""Telegram driver -- python-telegram-bot v22 long-poll.
 
-Bernstein historically embedded a ``python-telegram-bot`` long-poll
-loop inside its own process. That meant the bot died with bernstein
-and the same token could not be shared with sibling projects on the
-VPS. The new architecture is a standalone ``thisnotabot-router``
-service on the VPS which owns the webhook + dispatcher; each project
-ships its own :class:`thisnotabot_bridge.BridgeRouter` of slash
-commands and pushes notifications via :class:`BridgeNotifier`.
+Standard Telegram bot integration: get a bot API token from
+``@BotFather``, set ``BERNSTEIN_TG_TOKEN`` (or pass it via the chat
+sink config), and the bridge connects via long-poll. No external
+services. ``pip install 'bernstein[telegram]'`` pulls in the
+``python-telegram-bot`` library.
 
-This module is the bernstein-side glue:
+The library import is guarded so the module can always be imported --
+``python-telegram-bot`` is only required when :meth:`TelegramBridge.start`
+actually runs. That keeps ``bernstein chat serve --platform=discord``
+working for users who only installed the Discord extra.
 
-  * :meth:`on_command` registers handlers in the SDK's per-project
-    registry under ``project="bernstein"``. The standalone router
-    discovers them via :func:`thisnotabot_bridge.decorators.get_router`
-    and merges them into its dispatcher.
-  * :meth:`send_message` POSTs to ``/tg/notify`` with severity ``info``.
-  * :meth:`edit_message` and :meth:`push_approval` degrade gracefully
-    when the router does not yet support those primitives; the
-    transitional plan is documented in
-    ``research/telegram_bot_modernization_2026-05-09/design.md``.
-  * :meth:`start` and :meth:`stop` are essentially no-ops: the router
-    runs out-of-process so there is nothing for bernstein to listen on.
+Key behaviours:
 
-If :envvar:`BERNSTEIN_CHAT_USE_LEGACY` is set to ``1``/``true``/``yes``
-the factory in :mod:`bernstein.core.chat` reroutes back to the
-:class:`bernstein.core.chat.drivers._legacy_telegram.TelegramBridge`
-long-poll driver so an operator can fall back without code changes.
+  * **Slash commands.** Handlers registered via :meth:`on_command` are
+    routed based on the leading ``/word`` in each message. The raw
+    :class:`~bernstein.core.chat.bridge.ChatMessage` is forwarded so
+    handlers can read the remaining argv.
+  * **Approval buttons.** :meth:`push_approval` renders an
+    ``InlineKeyboardMarkup`` with two callback buttons. The callback
+    data encodes ``approve:<id>`` / ``reject:<id>`` so the driver can
+    resolve the decision back to the pending approval without any
+    extra state.
+  * **Edit throttle.** :meth:`edit_message` is debounced to one edit
+    per thread per 500ms. Telegram rate-limits bot edits aggressively
+    and a streaming agent will otherwise melt the channel.
 """
 
 from __future__ import annotations
 
+import asyncio
 import importlib
 import logging
-import os
-from typing import TYPE_CHECKING, Any
+import time
+from dataclasses import dataclass, field
+from typing import Any
 
 from bernstein.core.chat.bridge import (
     BridgeProtocol,
@@ -43,335 +44,302 @@ from bernstein.core.chat.bridge import (
     PendingApproval,
 )
 
-if TYPE_CHECKING:
-    from collections.abc import Awaitable, Callable
-
-__all__ = [
-    "BRIDGE_PROJECT",
-    "BridgeDependencyError",
-    "TelegramBridge",
-    "is_legacy_mode_enabled",
-]
+__all__ = ["EDIT_THROTTLE_S", "TelegramBridge", "TelegramDependencyError"]
 
 logger = logging.getLogger(__name__)
 
-#: Project namespace registered with the SDK's ``@command`` decorator.
-BRIDGE_PROJECT = "bernstein"
-
-#: Env var that toggles the legacy long-poll driver as a hard fallback.
-LEGACY_ENV_FLAG = "BERNSTEIN_CHAT_USE_LEGACY"
+#: Minimum seconds between consecutive edits to the same message id.
+EDIT_THROTTLE_S: float = 0.5
 
 
-class BridgeDependencyError(RuntimeError):
-    """Raised when ``thisnotabot-bridge`` is not installed."""
+class TelegramDependencyError(RuntimeError):
+    """Raised when ``python-telegram-bot`` is not installed."""
 
 
-def is_legacy_mode_enabled() -> bool:
-    """Return True when the operator opted into the legacy long-poll path.
+@dataclass(slots=True)
+class _EditState:
+    """Per-message debouncing bookkeeping.
 
-    Truthy values: ``1``, ``true``, ``yes``, ``on`` (case-insensitive).
-    Anything else (including an unset variable) keeps the modern bridge
-    path active.
+    Attributes:
+        last_edit_ts: Monotonic timestamp of the last successful flush.
+        pending_text: Latest body awaiting flush. Empty string means no
+            pending write.
+        task: Scheduled flush coroutine, if any.
     """
-    raw = os.environ.get(LEGACY_ENV_FLAG, "")
-    return raw.strip().lower() in {"1", "true", "yes", "on"}
+
+    last_edit_ts: float = 0.0
+    pending_text: str = ""
+    task: asyncio.Task[None] | None = field(default=None, repr=False)
 
 
 class TelegramBridge(BridgeProtocol):
-    """Telegram driver delegating to the ``thisnotabot-router`` service.
-
-    The class still implements :class:`BridgeProtocol` so existing
-    callers (the CLI, the ``ChatSession`` glue, the notification sink)
-    keep working unchanged. The transport is HTTP-out only -- inbound
-    Telegram updates are received by the standalone router service and
-    dispatched into bernstein via the aiogram :class:`Router` returned
-    by :meth:`thisnotabot_bridge.BridgeRouter.to_aiogram`.
-    """
+    """Telegram implementation of :class:`BridgeProtocol`."""
 
     platform: str = "telegram"
 
-    def __init__(self, token: str = "", *, project: str | None = None) -> None:
-        """Build a bridge bound to the bernstein project namespace.
+    def __init__(self, token: str) -> None:
+        """Create a bridge bound to a Telegram bot ``token``.
 
-        ``token`` is accepted for API parity with the legacy driver but
-        ignored: the standalone router holds the actual Telegram token.
-        Local code only needs the per-project ``X-Notify-Secret`` from
-        the env -- :class:`thisnotabot_bridge.BridgeNotifier` reads it
-        on demand.
+        The token is captured eagerly but no network I/O happens until
+        :meth:`start` is called.
         """
-        # Token is accepted but never used here -- preserved purely for
-        # call-site compatibility with the legacy ``TelegramBridge``.
+        if not token:
+            raise ValueError("Telegram bot token must be non-empty.")
         self._token = token
-        self._project = project or BRIDGE_PROJECT
         self._command_handlers: dict[str, CommandHandler] = {}
         self._button_handler: ButtonHandler | None = None
-        self._notifier: Any = None
-        self._router: Any = None
-        self._sdk: Any = None
-
-    # ------------------------------------------------------------------
-    # SDK plumbing
-    # ------------------------------------------------------------------
-
-    def _ensure_sdk(self) -> Any:
-        """Import the SDK lazily; raise a useful error if it's missing.
-
-        The bridge is the entire point of the shim, so the dependency
-        is mandatory in this code path. Tests that monkeypatch the SDK
-        replace the ``sys.modules`` entry before construction.
-        """
-        if self._sdk is not None:
-            return self._sdk
-        try:
-            self._sdk = importlib.import_module("thisnotabot_bridge")
-        except ImportError as exc:
-            raise BridgeDependencyError(
-                "thisnotabot-bridge is not installed. "
-                "Install with: pip install -e /path/to/thisnotabot, "
-                "or set BERNSTEIN_CHAT_USE_LEGACY=1 to use the long-poll driver.",
-            ) from exc
-        return self._sdk
-
-    def _ensure_notifier(self) -> Any:
-        """Build (lazily) the :class:`BridgeNotifier` HTTP client."""
-        if self._notifier is not None:
-            return self._notifier
-        sdk = self._ensure_sdk()
-        notifier_cls: Any = sdk.BridgeNotifier
-        self._notifier = notifier_cls.from_env(project=self._project)
-        return self._notifier
-
-    def _ensure_router(self) -> Any:
-        """Get-or-create the SDK's per-project :class:`BridgeRouter`."""
-        if self._router is not None:
-            return self._router
-        sdk = self._ensure_sdk()
-        get_router: Any = sdk.decorators.get_router  # public via sub-module
-        self._router = get_router(self._project)
-        return self._router
+        self._app: Any = None
+        self._edit_state: dict[str, _EditState] = {}
+        self._edit_lock = asyncio.Lock()
 
     # ------------------------------------------------------------------
     # Registration
     # ------------------------------------------------------------------
 
     def on_command(self, name: str, handler: CommandHandler) -> None:
-        """Register ``handler`` for ``/<name>`` on the bernstein router.
-
-        The SDK's ``@command`` decorator drops a :class:`CommandSpec`
-        into a process-global registry keyed by project name. We bypass
-        the decorator surface and write directly to that registry so
-        the chat-session keeps owning lifecycle (the ``ChatSession``
-        instantiates fresh handlers per ``serve`` invocation; we don't
-        want module-level decoration to leak across runs).
-        """
-        clean = name.lstrip("/")
-        self._command_handlers[clean] = handler
-        sdk = self._ensure_sdk()
-        spec_cls: Any = sdk.decorators.CommandSpec
-        spec = spec_cls(
-            name=clean,
-            project=self._project,
-            desc=_describe(clean),
-            handler=_make_aiogram_adapter(handler),
-        )
-        router = self._ensure_router()
-        # ``add`` is idempotent in the trivial sense -- duplicate names
-        # on the same project are intentionally appended; the router
-        # service de-duplicates on ``setMyCommands`` registration.
-        router.add(spec)
+        """Register ``handler`` for the slash command ``/<name>``."""
+        self._command_handlers[name.lstrip("/")] = handler
 
     def on_button(self, handler: ButtonHandler) -> None:
-        """Record the approval button handler.
-
-        The current SDK does not yet expose inline-keyboard callback
-        plumbing, so the handler is captured for future use only.
-        Slash-command approve / reject (the bernstein default for chat
-        approvals) work via :meth:`on_command` exactly as before.
-        """
+        """Register the single approve/reject callback."""
         self._button_handler = handler
 
     # ------------------------------------------------------------------
-    # Lifecycle (router lives out-of-process; these are mostly no-ops)
+    # Lifecycle
     # ------------------------------------------------------------------
 
     async def start(self) -> None:
-        """Validate the SDK and notifier are reachable.
+        """Connect to Telegram and begin polling for updates."""
+        tg_ext: Any = _import_telegram_ext()
+        self._app = tg_ext.Application.builder().token(self._token).build()
 
-        We never spawn a long-poll loop here -- the standalone
-        ``thisnotabot-router`` does that. We do, however, eagerly
-        materialise the notifier so any misconfiguration (missing
-        ``TELEGRAM_THISNOTABOT_INTERNAL_TOKEN`` etc.) surfaces at boot
-        rather than the first ``send_message``.
-        """
-        self._ensure_notifier()
-        self._ensure_router()
-        logger.info(
-            "telegram-bridge: started in router-delegate mode (project=%s, commands=%s)",
-            self._project,
-            sorted(self._command_handlers),
-        )
+        # Install every registered slash command.
+        tg_cmd_handler_cls: Any = tg_ext.CommandHandler
+        for name in self._command_handlers:
+            self._app.add_handler(
+                tg_cmd_handler_cls(name, self._dispatch_command),
+            )
+
+        cb_cls: Any = tg_ext.CallbackQueryHandler
+        self._app.add_handler(cb_cls(self._dispatch_button))
+
+        await self._app.initialize()
+        await self._app.start()
+        updater = getattr(self._app, "updater", None)
+        if updater is not None:
+            await updater.start_polling()
 
     async def stop(self) -> None:
-        """Drop in-process state. The router service keeps running."""
-        self._notifier = None
-        self._router = None
+        """Flush pending edits and disconnect cleanly."""
+        # Cancel any in-flight throttled flushes so the loop can exit.
+        async with self._edit_lock:
+            for state in self._edit_state.values():
+                task = state.task
+                if task is not None and not task.done():
+                    task.cancel()
+            self._edit_state.clear()
+
+        if self._app is None:
+            return
+        updater = getattr(self._app, "updater", None)
+        if updater is not None:
+            await updater.stop()
+        await self._app.stop()
+        await self._app.shutdown()
+        self._app = None
 
     # ------------------------------------------------------------------
     # Outbound primitives
     # ------------------------------------------------------------------
 
     async def send_message(self, thread_id: str, text: str) -> str:
-        """Push ``text`` to ``thread_id`` via the router's notify endpoint.
-
-        Returns the empty string -- the router does not echo the
-        Telegram message id back to the SDK. Callers that need a real
-        id should keep using the legacy driver until the router learns
-        to forward send results (planned).
-        """
-        notifier = self._ensure_notifier()
-        chat_id = _coerce_chat_id(thread_id)
-        sdk = self._ensure_sdk()
-        severity = sdk.NotifySeverity.INFO
-        title, body = _split_title_body(text)
-        try:
-            await notifier.notify(
-                title,
-                body,
-                severity=severity,
-                chat_id_override=chat_id,
-            )
-        except Exception as exc:  # pragma: no cover - HTTP-only path
-            logger.warning(
-                "telegram-bridge: notify failed thread=%s err=%s",
-                thread_id,
-                exc,
-            )
-        return ""
+        """Post ``text`` to ``thread_id`` and return the new message id."""
+        app = self._require_app()
+        sent = await app.bot.send_message(chat_id=_to_chat_id(thread_id), text=text)
+        return str(sent.message_id)
 
     async def edit_message(self, thread_id: str, message_id: str, text: str) -> None:
-        """No-op shim for streaming edits.
+        """Edit ``message_id`` in ``thread_id``, debounced to 2Hz per message.
 
-        The router does not yet expose ``editMessageText`` over the
-        notify protocol. Streaming agents that called ``edit_message``
-        on the legacy driver will instead see a sequence of fresh
-        notifications until the bridge gains an edit endpoint.
+        Rapid successive calls to the same ``message_id`` collapse into
+        a single deferred write, guaranteeing at most one Telegram API
+        call every :data:`EDIT_THROTTLE_S` seconds. The most recently
+        requested body always wins.
         """
-        # Best-effort fallback: send a *new* message so the operator
-        # still sees progress -- otherwise the chat goes silent.
-        del message_id  # unused on this path
-        await self.send_message(thread_id, text)
+        key = f"{thread_id}:{message_id}"
+        now = time.monotonic()
+        async with self._edit_lock:
+            state = self._edit_state.setdefault(key, _EditState())
+            state.pending_text = text
+            elapsed = now - state.last_edit_ts
+            if elapsed >= EDIT_THROTTLE_S and (state.task is None or state.task.done()):
+                state.last_edit_ts = now
+                body = state.pending_text
+                state.pending_text = ""
+                await self._flush_edit(thread_id, message_id, body)
+                return
+            if state.task is None or state.task.done():
+                delay = max(0.0, EDIT_THROTTLE_S - elapsed)
+                state.task = asyncio.create_task(
+                    self._deferred_flush(thread_id, message_id, delay, key),
+                )
 
     async def push_approval(self, approval: PendingApproval) -> str:
-        """Render an approval as a notify message with manual instructions.
+        """Render an inline approve/reject card for ``approval``."""
+        tg: Any = _import_telegram()
+        keyboard_cls: Any = tg.InlineKeyboardMarkup
+        button_cls: Any = tg.InlineKeyboardButton
+        keyboard = keyboard_cls(
+            [
+                [
+                    button_cls("Approve", callback_data=f"approve:{approval.approval_id}"),
+                    button_cls("Reject", callback_data=f"reject:{approval.approval_id}"),
+                ],
+            ],
+        )
+        app = self._require_app()
+        sent = await app.bot.send_message(
+            chat_id=_to_chat_id(approval.thread_id),
+            text=f"{approval.title}\n\n{approval.body}",
+            reply_markup=keyboard,
+        )
+        return str(sent.message_id)
 
-        The SDK does not expose an inline-keyboard primitive yet, so we
-        send a warning-severity notification asking the operator to
-        type ``/approve`` or ``/reject``. The slash-command path is
-        already wired through :meth:`on_command` and reads the same
-        pending-approval directory the legacy buttons used.
-        """
-        sdk = self._ensure_sdk()
-        notifier = self._ensure_notifier()
-        chat_id = _coerce_chat_id(approval.thread_id)
-        body = f"{approval.body}\n\nApproval id: {approval.approval_id}\nReply with /approve or /reject to resolve."
-        try:
-            await notifier.notify(
-                approval.title,
-                body,
-                severity=sdk.NotifySeverity.WARNING,
-                chat_id_override=chat_id,
-            )
-        except Exception as exc:  # pragma: no cover - HTTP-only path
-            logger.warning(
-                "telegram-bridge: push_approval failed approval=%s err=%s",
-                approval.approval_id,
-                exc,
-            )
-        return ""
+    # ------------------------------------------------------------------
+    # Inbound dispatch -- wired into python-telegram-bot handlers.
+    # ------------------------------------------------------------------
 
-
-# ---------------------------------------------------------------------------
-# Helpers
-# ---------------------------------------------------------------------------
-
-
-_COMMAND_DESCRIPTIONS: dict[str, str] = {
-    "run": 'Start a new agent task: /run "<goal>"',
-    "status": "Show the current chat-thread session.",
-    "approve": "Approve the oldest pending tool call.",
-    "reject": "Reject the oldest pending tool call.",
-    "switch": "Re-dispatch the current goal: /switch <adapter>",
-    "stop": "Stop the active session in this thread.",
-    "handoff": "Issue or claim a cross-surface session handoff token.",
-}
-
-
-def _describe(name: str) -> str:
-    """Lookup a human-readable description for ``setMyCommands``."""
-    return _COMMAND_DESCRIPTIONS.get(name, f"bernstein /{name}")
-
-
-def _coerce_chat_id(thread_id: str) -> int | None:
-    """Convert the bernstein string thread-id to the int the router needs."""
-    if not thread_id:
-        return None
-    try:
-        return int(thread_id)
-    except (TypeError, ValueError):
-        return None
-
-
-def _split_title_body(text: str) -> tuple[str, str]:
-    """Split a multiline payload into ``(title, body)`` for the notify card.
-
-    The title cap is 256 chars (the router's pydantic constraint); the
-    rest goes into the body. Single-line inputs use the whole text as
-    the title with an empty body.
-    """
-    if not text:
-        return ("(empty)", "")
-    head, _, tail = text.partition("\n")
-    title = head.strip()[:256] or "(empty)"
-    body = tail.strip()
-    if not head.strip() and tail.strip():
-        # Pathological case: blank first line.
-        title = tail.strip().splitlines()[0][:256]
-        body = "\n".join(tail.strip().splitlines()[1:])
-    return title, body
-
-
-def _make_aiogram_adapter(
-    handler: CommandHandler,
-) -> Callable[..., Awaitable[None]]:
-    """Wrap a bernstein :class:`CommandHandler` for aiogram dispatch.
-
-    The router service materialises the SDK's :class:`BridgeRouter`
-    via :meth:`thisnotabot_bridge.BridgeRouter.to_aiogram`, which calls
-    ``handler(message)`` with an ``aiogram.types.Message``. We adapt
-    that into the platform-agnostic :class:`ChatMessage` bernstein's
-    handlers expect.
-    """
-
-    async def _aiogram_handler(message: Any) -> None:
-        text = str(getattr(message, "text", "") or "")
-        chat = getattr(message, "chat", None)
-        chat_id = str(getattr(chat, "id", "") or "")
-        from_user = getattr(message, "from_user", None)
-        user_id = str(getattr(from_user, "id", "") or "")
-        message_id = str(getattr(message, "message_id", "") or "")
-        # aiogram strips the leading ``/cmd`` token via ``Command`` filter
-        # but the raw text still contains it; mirror legacy parsing.
-        parts = text.split() if text else []
-        args = parts[1:]
+    async def _dispatch_command(self, update: Any, _context: Any) -> None:
+        """Route a ``/command`` to the matching registered handler."""
+        message = _extract_message(update)
+        if message is None:
+            return
+        text: str = str(message.text or "")
+        if not text.startswith("/"):
+            return
+        parts: list[str] = [str(p) for p in text.split()]
+        name = parts[0][1:].split("@", 1)[0]  # strip @botname suffix
+        handler = self._command_handlers.get(name)
+        if handler is None:
+            return
+        chat_id = str(getattr(message.chat, "id", "") or "")
+        user = getattr(message, "from_user", None)
+        user_id = str(getattr(user, "id", "") or "")
         await handler(
             ChatMessage(
                 thread_id=chat_id,
                 user_id=user_id,
                 text=text,
-                message_id=message_id,
-                args=args,
-                raw=message,
+                message_id=str(getattr(message, "message_id", "") or ""),
+                args=parts[1:],
+                raw=update,
             ),
         )
 
-    return _aiogram_handler
+    async def _dispatch_button(self, update: Any, _context: Any) -> None:
+        """Route an ``InlineKeyboardButton`` press to :attr:`_button_handler`."""
+        if self._button_handler is None:
+            return
+        query = getattr(update, "callback_query", None)
+        if query is None:
+            return
+        data = str(getattr(query, "data", "") or "")
+        if ":" not in data:
+            return
+        decision, approval_id = data.split(":", 1)
+        if decision not in {"approve", "reject"}:
+            return
+        answer: Any = getattr(query, "answer", None)
+        if callable(answer):
+            result: Any = answer()
+            if hasattr(result, "__await__"):
+                await result
+        message = getattr(query, "message", None)
+        chat_id = str(getattr(getattr(message, "chat", None), "id", "") or "")
+        await self._button_handler(chat_id, approval_id, decision)
+
+    # ------------------------------------------------------------------
+    # Throttle internals
+    # ------------------------------------------------------------------
+
+    async def _deferred_flush(
+        self,
+        thread_id: str,
+        message_id: str,
+        delay: float,
+        key: str,
+    ) -> None:
+        """Sleep ``delay`` then flush the pending body for ``key``."""
+        try:
+            await asyncio.sleep(delay)
+        except asyncio.CancelledError:
+            # Propagate cancellation after releasing local state; swallowing
+            # CancelledError silently would desync the asyncio task tree
+            # (Sonar python:S7497).
+            raise
+        async with self._edit_lock:
+            state = self._edit_state.get(key)
+            if state is None or not state.pending_text:
+                return
+            body = state.pending_text
+            state.pending_text = ""
+            state.last_edit_ts = time.monotonic()
+        await self._flush_edit(thread_id, message_id, body)
+
+    async def _flush_edit(self, thread_id: str, message_id: str, text: str) -> None:
+        """Issue the actual ``edit_message_text`` API call."""
+        app = self._require_app()
+        try:
+            await app.bot.edit_message_text(
+                chat_id=_to_chat_id(thread_id),
+                message_id=int(message_id),
+                text=text,
+            )
+        except Exception as exc:  # pragma: no cover - network-only path.
+            logger.warning("telegram edit failed for %s:%s: %s", thread_id, message_id, exc)
+
+    def _require_app(self) -> Any:
+        if self._app is None:
+            raise RuntimeError(
+                "TelegramBridge is not started; call await bridge.start() first.",
+            )
+        return self._app
+
+
+# ---------------------------------------------------------------------------
+# Import helpers -- keep the SDK optional.
+# ---------------------------------------------------------------------------
+
+
+def _import_telegram() -> Any:
+    try:
+        return importlib.import_module("telegram")
+    except ImportError as exc:  # pragma: no cover - exercised via tests with stubbed module.
+        raise TelegramDependencyError(
+            "python-telegram-bot is not installed. Install with: pip install 'bernstein[telegram]'",
+        ) from exc
+
+
+def _import_telegram_ext() -> Any:
+    try:
+        return importlib.import_module("telegram.ext")
+    except ImportError as exc:  # pragma: no cover - exercised via tests with stubbed module.
+        raise TelegramDependencyError(
+            "python-telegram-bot is not installed. Install with: pip install 'bernstein[telegram]'",
+        ) from exc
+
+
+def _extract_message(update: Any) -> Any:
+    """Return ``update.message`` or ``update.edited_message`` if present."""
+    message = getattr(update, "message", None)
+    if message is not None:
+        return message
+    return getattr(update, "edited_message", None)
+
+
+def _to_chat_id(thread_id: str) -> int | str:
+    """Coerce ``thread_id`` to ``int`` for Telegram's integer chat ids."""
+    try:
+        return int(thread_id)
+    except (TypeError, ValueError):
+        return thread_id
